Бахарева Антонина Задание04

+++ task04/src/com/example/task04/Point.java
@@ -1,8 +1,8 @@
 package com.example.task04;
 
 public class Point {
-    final int x;
-    final int y;
+    final double x;
+    final double y;
 
     public Point(int x, int y) {
         this.x = x;
@@ -10,6 +10,6 @@ public class Point {
     }
 
     public String toString() {
-        return String.format("(%d, %d)", x, y);
+        return String.format("(%g, %g)", x, y);
     }
 }
